Add S3 Storage module
Add an implementation of the storage interface that uses an S3 bucket to
store the diff, "in-progrress" indicator, and supports "marking"
something in-progress. This allows the POST endpoint to check to see if
a diff already exists in S3 for given start and end times and to mark a
diff in-progress once it queses the diff for creation. It also allows the
GET endpoint to fetch a diff from S3 and supports proper handling if the
diff is in-progress.