Replace DOT library to reduce memory pressure
Replace full loading and parsing with data streaming and string
manipulation. This increases risk that faulting input will crash the
program but greatly reduces the memory consumption required to generate
the diff. A full disclaimer is added to the Diff method.

The rough results of this are a 70% decrease in runtime and 70% decrease
in allocations. The overall space allocated during the benchmark is
reduced by 40%. There are more optimization we can make from here if
needed.